+++ scripts/git-subtree/custom-block/pull.ps1
@@ -1,0 +1,4 @@
+git subtree pull `
+    -P workspace/plugins/custom-block `
+    https://github.com/Zuoqiu-Yingyi/siyuan-plugin-custom-block.git `
+    dev